Make protein quest runnable from pyodide/wasm via marimo + make distributed/psutil non-dep for wasm
